move request id to separate file and use it as a key in protocol context result handlers
